Claude: Move request schemas into nr-common.
