+++ nr-common/mod.ts
@@ -13,5 +13,6 @@ export * from "./src/30398.schema.ts";
 export * from "./src/event.schema.ts";
 export * from "./src/isValidEvent.ts";
 export * from "./src/notifications.schema.ts";
+export * from "./src/nrBridge.schema.ts";
 export * from "./src/plusCode.schema.ts";
 export * from "./src/getAuthorFromEvent.ts";
